news/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
defusedxml==0.7.1
Django==3.2.15
django-adminlte3==0.1.6
django-filter==21.1
djangorestframework==3.12.4
drf-extra-fields==3.4.0
flake8==5.0.4
flake8-broken-line==0.5.0
flake8-isort==4.2.0
flake8-plugin-utils==1.3.2
flake8-return==1.1.3
idna==3.3
importlib-metadata==1.7.0
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
mysql-connector-python==8.0.30
mysqlclient==2.1.0
pep8-naming==0.13.2
Pillow==9.2.0
protobuf==3.20.1
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.2.1
reportlab==3.6.11
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.11
zipp==3.8.1
